# feat:  compute leaf hashes in the Push method