--- a/hasher_test.go
+++ b/hasher_test.go
@@ -643,5 +643,4 @@ func Test_MustHashLeaf_panic(t *testing.T) {
 			}
 		})
 	}
-
 }