--- a/proof_test.go
+++ b/proof_test.go
@@ -48,7 +48,6 @@ func TestProof_VerifyNamespace_False(t *testing.T) {
 	leafIndex := 3
 	inclusionProofOfLeafIndex, err := n.buildRangeProof(leafIndex, leafIndex+1)
 	require.NoError(t, err)
-	require.NoError(t, n.computeLeafHashesIfNecessary())
 	leafHash := n.leafHashes[leafIndex] // the only data item with namespace ID = 2 in the constructed tree is at index 3
 	invalidAbsenceProof := NewAbsenceProof(leafIndex, leafIndex+1, inclusionProofOfLeafIndex, leafHash, false)