@staheri14 2023-04-03
Please find the benchmark comparison below. It's worth noting that the change introduced by this PR doesn't affect the [`ComputeRoot()` benchmark test](https://github.com/celestiaorg/nmt/blob/a20a72346da3ca2543c266a6d70fca1592f90b7f/nmt_test.go#L683) result. This is because the time measurement captures the total time taken to push `N` items and then compute the root. Before this change, the items were hashed inside the `Root()` method, whereas now they are hashed within the `Push()` method. As a result, the total number of hashes remains unchanged, and therefore the performance for this particular benchmark test does not improve.

```
goos: darwin
goarch: arm64
pkg: github.com/celestiaorg/nmt
                          │   old.txt   │              new.txt               │
                          │   sec/op    │   sec/op     vs base               │
ComputeRoot/64-leaves-10    36.28µ ± 1%   38.10µ ± 0%  +5.00% (p=0.000 n=10)
ComputeRoot/128-leaves-10   72.23µ ± 0%   75.29µ ± 0%  +4.24% (p=0.000 n=10)
ComputeRoot/256-leaves-10   146.4µ ± 1%   152.7µ ± 1%  +4.30% (p=0.000 n=10)
geomean                     72.66µ        75.94µ       +4.51%
```